requirements.txt
gdown==5.2.0
keras==3.8.0
matplotlib==3.9.2
mpmath==1.3.0
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.3
pillow==10.4.0
scipy==1.15.1
tensorboard==2.16.2
tensorflow==2.16.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
ultralytics==8.3.73
utils==1.0.2